Add GetById and GetAll endpoints for OS entity
Creates the controller and service methods to implement the "Get By Id" and "Get All" endpoints for the "Ordem" entity.

It also makes some major changes in naming conventions, the main one being the change of the name associated to the "Ordem" entity, from "OrdemdeServico" to just "Ordem."

The type of the class property "serialVersionUID" was also altered from "Long" to "long ", which made a warning of unused variable go away.